+++ pkg/ml/nn/param.go
@@ -291,7 +291,7 @@ func (r *param) ID() int64 {
 }
 
 // TimeStep returns always 0 since the "pure" parameter is not associated with any graph.
-func (r *param) TimeStep() int64 {
+func (r *param) TimeStep() int {
 	return 0
 }
